pyproject.toml
[project]
name = "h4sg25-cdl-challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dspy>=2.6.16",
    "fastexcel>=0.13.0",
    "jupyter>=1.1.1",
    "polars>=1.26.0",
    "pytest>=8.3.5",
    "python-levenshtein>=0.27.1",
    "seaborn>=0.13.2",
    "spacy>=3.8.5",
    "squarify>=0.4.4",
]


[tool.uv.sources]
funding_crawler = { path = "helpers" }
